+++ app/src/main/java/com/hongul/filq/FliQApp.kt
@@ -3,7 +3,7 @@ package com.hongul.filq
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import com.hongul.filq.ui.NavigationGraph
+import com.hongul.filq.ui.navigation.NavigationGraph
 
 @Composable
 fun FliQApp(navController: NavHostController = rememberNavController()) {
