+++ app/src/main/java/com/hongul/filq/MainActivity.kt
@@ -7,8 +7,6 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.safeDrawingPadding
 import androidx.compose.ui.Modifier
-import androidx.navigation.compose.rememberNavController
-import com.hongul.filq.ui.NavigationGraph
 import com.hongul.filq.ui.theme.FilQTheme
 
 class MainActivity : ComponentActivity() {
